svn merge -r 29034:29035 svn://chef/work/src .
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/branch/prod/release-9.0/work/src@29037 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ firewall/impl/com/untangle/node/firewall/EventHandler.java
@@ -193,8 +193,8 @@ class EventHandler extends AbstractEventHandler
     }
 
     protected FirewallMatcher findMatchingRule( Protocol protocol,
-                                                byte clientIntf, InetAddress clientAddr, int clientPort,
-                                                byte serverIntf, InetAddress serverAddr, int serverPort)
+                                                int clientIntf, InetAddress clientAddr, int clientPort,
+                                                int serverIntf, InetAddress serverAddr, int serverPort)
     {
         for (FirewallMatcher matcher : firewallRuleList) {
